sentry - mod test_util - setup_dummy_app helper fn
