+++ sentry/src/middleware/campaign.rs
@@ -37,57 +37,18 @@ impl<A: Adapter + 'static> Middleware<A> for CampaignLoad {
 
 #[cfg(test)]
 mod test {
-    use adapter::DummyAdapter;
     use primitives::{
-        adapter::DummyAdapterOptions,
-        config::configuration,
-        util::tests::{
-            discard_logger,
-            prep_db::{DUMMY_CAMPAIGN, IDS},
-        },
+        util::tests::prep_db::{DUMMY_CAMPAIGN, IDS},
         Campaign,
     };
 
-    use crate::db::{
-        insert_campaign,
-        redis_pool::TESTS_POOL,
-        tests_postgres::{setup_test_migrations, DATABASE_POOL},
-    };
+    use crate::{db::insert_campaign, test_util::setup_dummy_app};
 
     use super::*;
 
-    async fn setup_app() -> Application<DummyAdapter> {
-        let config = configuration("development", None).expect("Should get Config");
-        let adapter = DummyAdapter::init(
-            DummyAdapterOptions {
-                dummy_identity: IDS["leader"],
-                dummy_auth: Default::default(),
-                dummy_auth_tokens: Default::default(),
-            },
-            &config,
-        );
-
-        let redis = TESTS_POOL.get().await.expect("Should return Object");
-        let database = DATABASE_POOL.get().await.expect("Should get a DB pool");
-
-        setup_test_migrations(database.pool.clone())
-            .await
-            .expect("Migrations should succeed");
-
-        let app = Application::new(
-            adapter,
-            config,
-            discard_logger(),
-            redis.connection.clone(),
-            database.pool.clone(),
-        );
-
-        app
-    }
-
     #[tokio::test]
     async fn campaign_loading() {
-        let app = setup_app().await;
+        let app = setup_dummy_app().await;
 
         let build_request = |params: RouteParams| {
             Request::builder()
